# onboarding/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Interactive onboarding platform: tours, immutable versions, resolve and analytics.",
        "title": "Cluer Onboarding",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/apps": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apps"
                ],
                "summary": "List consumer applications",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.App"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Issues the public key that goes into the SDK script tag",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Apps"
                ],
                "summary": "Register a consumer application",
                "parameters": [
                    {
                        "description": "Application",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateAppRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.App"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
        "/events": {
            "post": {
                "description": "Idempotent on event_key, so a retry after a dropped connection cannot double the metrics",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Runtime"
                ],
                "summary": "Accept a batch of events",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application public key",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Event batch",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.EventBatchRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/response.EventBatchResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/resolve": {
            "post": {
                "description": "The single request the SDK makes on init. Returns at most one tour, whole.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Runtime"
                ],
                "summary": "Resolve a tour for a subject",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Application public key",
                        "name": "X-App-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Resolve context",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResolveResult"
                        }
                    },
                    "204": {
                        "description": "Nothing to show"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/tours/{tourId}/analytics": {
            "get": {
                "description": "Per-hint funnel for one version over one period",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Tour funnel and totals",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Tour ID",
                        "name": "tourId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "Version ID, defaults to published",
                        "name": "versionId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Period start, RFC3339",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Period end, RFC3339",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Analytics"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errs.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "errs.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Shorthand error",
                    "type": "string"
                },
                "message": {
                    "description": "Pretty error message",
                    "type": "string"
                }
            }
        },
...
        "request.CreateAppRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "allowed_origins": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "request.EventBatchRequest": {
            "type": "object",
            "required": [
                "events",
                "session_id",
                "subject_id"
            ],
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.EventItem"
                    }
                },
                "session_id": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                }
            }
        },
        "request.EventItem": {
            "type": "object",
            "required": [
                "event_key",
                "type"
            ],
            "properties": {
                "event_key": {
                    "type": "string"
                },
                "hint_id": {
                    "type": "string"
                },
                "occurred_at": {
                    "type": "string"
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "tour_id": {
                    "type": "string"
                },
                "tour_version_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "tour_started",
                        "hint_shown",
                        "hint_completed",
                        "hint_skipped",
                        "selector_missing",
                        "tour_completed",
                        "tour_dismissed",
                        "goal_reached",
                        "custom"
                    ]
                }
            }
        },
...
        "request.ResolveRequest": {
            "type": "object",
            "required": [
                "subject_id"
            ],
            "properties": {
                "props": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "session_id": {
                    "type": "string"
                },
                "subject_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "response.Analytics": {
            "type": "object",
            "properties": {
                "broken_selectors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "from": {
                    "type": "string",
                    "format": "date-time"
                },
                "funnel": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.FunnelStep"
                    }
                },
                "to": {
                    "type": "string",
                    "format": "date-time"
                },
                "totals": {
                    "$ref": "#/definitions/response.AnalyticsTotals"
                },
                "tour_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "tour_version_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "version": {
                    "type": "integer",
                    "example": 7
                }
            }
        },
        "response.AnalyticsTotals": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer",
                    "example": 612
                },
                "completion_rate": {
                    "type": "number",
                    "example": 0.494
                },
                "dismissed": {
                    "type": "integer",
                    "example": 388
                },
                "goal_rate": {
                    "type": "number",
                    "example": 0.404
                },
                "goal_reached": {
                    "type": "integer",
                    "example": 501
                },
                "started": {
                    "type": "integer",
                    "example": 1240
                }
            }
        },
        "response.App": {
            "type": "object",
            "properties": {
                "allowed_origins": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "http://localhost:3000"
                    ]
                },
                "archived_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "id": {
                    "type": "string",
                    "format": "uuid"
                },
                "name": {
                    "type": "string",
                    "example": "Demo classifieds"
                },
                "public_key": {
                    "type": "string",
                    "example": "pk_demo_7f3a91c4b2e85d60"
                }
            }
        },
        "response.Audience": {
            "type": "object",
            "properties": {
                "max_shows": {
                    "type": "integer",
                    "example": 3
                },
                "only_new": {
                    "type": "boolean",
                    "example": false
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.AudienceRule"
                    }
                },
                "show_once": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "response.AudienceRule": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string",
                    "example": "is_premium"
                },
                "operator": {
                    "type": "string",
                    "enum": [
                        "exists",
                        "not_exists",
                        "eq",
                        "neq",
                        "gt",
                        "gte",
                        "lt",
                        "lte",
                        "contains",
                        "not_contains",
                        "starts_with",
                        "ends_with"
                    ],
                    "example": "eq"
                },
                "timeframe": {
                    "type": "string",
                    "enum": [
                        "1h",
                        "24h",
                        "7d",
                        "30d",
                        "all_time"
                    ],
                    "example": "7d"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "event_performed",
                        "page_visited",
                        "user_property"
                    ],
                    "example": "page_visited"
                },
                "value": {}
            }
        },
...
        "response.EventBatchResult": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "integer",
                    "example": 4
                },
                "duplicates": {
                    "type": "integer",
                    "example": 1
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rejected": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "response.FunnelStep": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "integer",
                    "example": 1105
                },
                "dropoff": {
                    "type": "number",
                    "example": 0.109
                },
                "hint_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "selector_missing": {
                    "type": "integer",
                    "example": 0
                },
                "shown": {
                    "type": "integer",
                    "example": 1240
                },
                "skipped": {
                    "type": "integer",
                    "example": 0
                },
                "step": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Начните здесь"
                }
            }
        },
...
        "response.Hint": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Здесь вы можете изменить параметры аккаунта"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "expected_input": {
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "01914f6a-8c9b-7a0b-9b0b-8c9b7a0b9b0c"
                },
                "input_placeholder": {
                    "type": "string",
                    "example": "Введите ваше имя"
                },
                "media_url": {
                    "type": "string",
                    "example": "https://cdn.example.com/onboarding/settings.mp4"
                },
                "page_path": {
                    "description": "Пусто — подсказка живёт на той же странице, где тур стартовал,\nто есть наследует target_path версии.",
                    "type": "string",
                    "example": "/dashboard"
                },
                "placement": {
                    "type": "string",
                    "enum": [
                        "top",
                        "bottom",
                        "center",
                        "left-top",
                        "right-top",
                        "left",
                        "right"
                    ],
                    "example": "bottom"
                },
                "required": {
                    "type": "boolean",
                    "example": false
                },
                "selector": {
                    "type": "string",
                    "example": "#settings-btn"
                },
                "spotlight": {
                    "type": "boolean",
                    "example": true
                },
                "step": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "example": "Нажмите кнопку настроек"
                },
                "tour_version_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "wait_for_selector": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "response.ResolveResult": {
            "type": "object",
            "properties": {
                "current_hint_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "tour": {
                    "$ref": "#/definitions/response.Tour"
                },
                "tour_version_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "version": {
                    "type": "integer",
                    "example": 7
                }
            }
        },
        "response.Tour": {
            "type": "object",
            "properties": {
                "app_id": {
                    "type": "string",
                    "format": "uuid"
                },
                "audience": {
                    "$ref": "#/definitions/response.Audience"
                },
                "created_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "description": {
                    "type": "string",
                    "example": "Тур по основным функциям дашборда"
                },
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "hints": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Hint"
                    }
                },
                "id": {
                    "type": "string",
                    "format": "uuid",
                    "example": "01914f6a-8c9b-7a0b-9b0b-8c9b7a0b9b0b"
                },
                "priority": {
                    "type": "integer",
                    "example": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "draft",
                        "published",
                        "archived"
                    ],
                    "example": "draft"
                },
                "target_path": {
                    "type": "string",
                    "example": "/dashboard"
                },
                "title": {
                    "type": "string",
                    "example": "Добро пожаловать в систему"
                },
                "trigger_config": {
                    "$ref": "#/definitions/response.TriggerConfig"
                },
                "trigger_type": {
                    "type": "string",
                    "enum": [
                        "on_load",
                        "delay",
                        "exit_intent",
                        "manual",
                        "scroll_depth",
                        "inactivity",
                        "element_visible"
                    ],
                    "example": "on_load"
                },
                "updated_at": {
                    "type": "string",
                    "format": "date-time"
                },
                "version": {
                    "type": "integer",
                    "example": 7
                },
                "version_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
...
        "response.TriggerConfig": {
            "type": "object",
            "properties": {
                "delay_ms": {
                    "type": "integer",
                    "example": 1500
                },
                "element_selector": {
                    "type": "string"
                },
                "inactivity_secs": {
                    "type": "integer",
                    "example": 30
                },
                "scroll_depth": {
                    "type": "integer",
                    "example": 50
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "OIDC id token: \"Bearer \u003cid_token\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}